+++ code/frontend/src/pages/SignupPage.jsx
@@ -3,7 +3,7 @@
 // Created by Natt 
 //Updated, integrated, and annotated by Natasya Liew
 
-import { Link, useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faEye, faEyeSlash } from '@fortawesome/free-solid-svg-icons'; // // Import password visibility toggle icons
